use ts-node and document script
