+++ global.d.ts
@@ -11,3 +11,6 @@ declare module '*.vert' {
 }
 
 declare module 'es6-promise-polyfill'
+declare module '@lerna/project'
+declare module '@lerna/batch-packages'
+declare module '@lerna/filter-packages'
